obs_next bug fix
For A2C based policies, compute_episodic_return requires obs_next. However, in multi-agent scenarios, obs_next was erroneously returning observations of the next agent, rather than returning the next observation of the current agent (issue 806). This caused incorrect computation of episodic returns, leading to poor convergence and issues with observations. Once such issue is that if different agents observe the exact same things, then the RL algorithm will work well, but if you switch the order of observations, the RL algorithm will stop working.

This issue was fixed by pre-proccessing the data differently in mapolicy. The order in which obs_next is assigned to agents was changed so that it matched with the correct order of the agents.