Refactoring: moved all trainers to the same module, added documentation
I also reduced some functional code duplication
between the offline and off-policy trainers

I believe this is useful since a lot of stuff was pure duplication,
in particular the inits and the docstrings of the trainers.
The implementations living in separate modules and
having significant code (which was 90% boilerplate) made
it difficult to spot the differences between these nearly
identical classes.

Having it all in one module and removing boilerplate should
result in a much clearer picture