[Fix&Enhance&Format] On README.md about poetry install (#1146)
- [x] I have added the correct label(s) to this Pull Request or linked
the relevant issue(s)
- [x] I have provided a description of the changes in this Pull Request
- [ ] I have added documentation for my changes and have listed relevant
changes in CHANGELOG.md
- [ ] If applicable, I have added tests to cover my changes.
- [ ] I have reformatted the code using `poe format` 
- [ ] I have checked style and types with `poe lint` and `poe
type-check`
- [ ] (Optional) I ran tests locally with `poe test` 
(or a subset of them with `poe test-reduced`) ,and they pass
- [ ] (Optional) I have tested that documentation builds correctly with
`poe doc-build`
---
1. [fix] First commit is **crucial**, otherwise `poetry` fails with `No
arguments expected for "install" command, got "envpool"`. It's clear in
[poetry-doc](https://python-poetry.org/docs/pyproject#extras) that quote
marks are needed.
2. [enh] Second commit is **optional**, I encountered this problem
myself (may be due to new to poetry), so I want to remind users (who may
be new to this, like me). However, I'm not sure if including this in the
main README is necessary, so it's up to you whether or not to merge it.
3. [fmt] Third commit is **recommended**, which is a minor but
meaningful formatting change (it was automatically formatted by my
VSCode plugin, and I didn't notice this at first).