Lower level access to hl world (#1187)
Gives the user the possibility to extract the `World` from a configured
`Experiment`. This required moving the initial rollout with random
actions (according to `start_timesteps`) from the `AgentFactory` to
`Experiment.run`. But since it was only ever used there, this change is
essentially backwards compatible

I also enhanced documentation. `__init__` docstrings were previously not
rendered and there were some formatting errors in them, hence the
changes there.

I also did some changes in the default-optim-factory creation in order
to not have to keep the defaults in `optim` and in `ExperimentBuilder`
in sync manually