Support computing custom scores and terminating/saving based on them in BaseTrainer (#1202)
This PR introduces a new concept into tianshou training: a `best_score`.
It is computed from the appropriate `Stats` instance and always added to
`InfoStats`.

## Breaking Changes:
- `InfoStats` has a new non-optional field `best_score`

## Background

Currently, tianshou uses the maximum average return to find the best
model. But sometimes it may not meet user needs, for example, the
average return only drops by 5%, but the standard deviation drops by
50%. The latter is generally considered to be more stable and better
than the former.