v2: Fix LaggedNetworkCollection.full_parameter_update forcing target network to eval mode
Achieve this by copying only the parameters (similar to Polyak update case) instead of
loading the other model's state dict, which is not strictly what the parameter update
was supposed to achieve.

Note: While forcing a target network to eval mode was appropriate in most cases,
RainbowDQN is a notable exception where a target network is required to be in training mode.